remove-collaborator api/ apis to separate file

+++ palinodes/static/palinodes/repository.js
@@ -31,7 +31,7 @@ document.addEventListener('DOMContentLoaded', function() {
         });
     });
 
-    let repositorypk = document.querySelector("#contents").dataset.pk;
+    const repositorypk = document.querySelector("#contents").dataset.pk;
     //LOAD CONTENTS
     loadDirectoryContents(repositorypk);
 
@@ -60,17 +60,35 @@ document.addEventListener('DOMContentLoaded', function() {
     var details = document.querySelector("#repository-details");
     var repositoryForm = details.querySelector("form");
 
+    //ADD COLLABORATORS
+    var addCollaboratorBtn = document.querySelector("#add-collaborator");
+    addCollaboratorBtn.onclick = searchUsers;
+
+    //REMOVE COLLABORATORS
+    details.querySelectorAll(".remove-collaborator").forEach(btn => {
+        btn.onclick = () => {
+            btn.parentElement.style.display = 'none';
+            removeCollaborator(btn.dataset.collaborator, repositorypk);
+        }
+    })
+
     //DELETE REPOSITORY
-    repositoryForm.querySelector("#delete-repository").onclick = () => {
-        deleteDirectory(repositorypk);
-        window.location.href = "/dashboard";
+    var deleteBtn = repositoryForm.querySelector("#delete-repository");
+    if(deleteBtn != null) {
+        deleteBtn.onclick = () => {
+            deleteDirectory(repositorypk);
+            window.location.href = "/dashboard";
+        }
     }
 
     //CANCEL EDIT
     repositoryForm.querySelector("#cancel-edit").onclick = () => {
         details.querySelector("#repository-settings").style.display = "flex";
         details.querySelector("#repository-details-details").style.display = 'flex';
-        details.querySelector("#repository-details-collaborators").style.display = 'block';
+        details.querySelector("#add-collaborator").style.display = "none";
+        details.querySelectorAll(".remove-collaborator").forEach(button => {
+            button.style.display = "none";
+        })
         details.querySelector("form").style.display = "none";
     }
 
@@ -78,16 +96,44 @@ document.addEventListener('DOMContentLoaded', function() {
     document.querySelector("#repository-settings").addEventListener('click', () => {
         details.querySelector("#repository-settings").style.display = "none";
         details.querySelector("#repository-details-details").style.display = "none";
-        details.querySelector("#repository-details-collaborators").style.display = 'none';
+        details.querySelector("#add-collaborator").style.display = "block";
+        details.querySelectorAll(".remove-collaborator").forEach(button => {
+            button.style.display = "block";
+        })
         
         repositoryForm.style.display = "block";
         
         repositoryForm.querySelector("#id_name").value = details.dataset.name;
         repositoryForm.querySelector("#id_description").value = details.dataset.description;
-        repositoryForm.querySelector("#id_collaborators").value = details.dataset.collaborators;
-        
     })
 });
+
+function removeCollaborator(collaboratorpk, repositorypk) {
+    //remove collaborator with api call
+    fetch(`/remove-collaborator/${repositorypk}`, {
+        method: 'post',
+        headers: {
+            'ContentType': "application/json",
+            'X-CSRFToken': csrftoken,
+        },
+        body: JSON.stringify({
+            'pk': collaboratorpk,
+        })
+    })
+    return
+}
+
+function searchUsers() {
+    //display search bar
+    //when the user starts typing, show the first _ results that contain the substring in the search bar
+        //detect typed event 
+            //get first ten users that have the value of the input as a substring
+                //display each user in a modal with an ADD button
+                // handle ADD click event
+                    //use a fetch call to add the collaborator
+
+    return
+}
 
 async function deleteDirectory(pk) {
     fetch("/delete-directory", {
